ISSUE: BugID: 456
PURPOSE: In Determinations > Determination Citations table, displays a designator's HUH Name.  If the designator doesn't have an HUH Name, displays "[no variant name]" to alert the user to add an HUH Name to the agent.

DESCRIPTION: Modified the botany default and manager view.xml files to use a new typedefsearch that uses a new dataobj formatter.  The original formatter searched for a different variant name, so I set the new one to search for HUH Name.