ISSUE: BugId: 448
PURPOSE: Fix/Generate hybrid full names according to Kanchi's requirements.  For genera, the fullname is xGenus.  For species, the full name is Genus xEpithet.  For taxa below species, append "notho" to the rank specifier.  The x in this comment is in reality the multiplication symbol.

Configure and run src.edu.harvard.huh.specify.teststuff.HybridFullNamesFixer to fix the existing hybrid names.

DESCRIPTION: Let Specify generate the full names.  If a taxa is a hybrid, use my rules in HUHTaxonBusRules.  This is compatible with Specify's visual tree drag and drop editor, as it calls HUHTaxonBusRules for each change.