ISSUE: BugID 495 (Tree by Parent branch)
PURPOSE: For trees, determining whether to enable the delete button was slow.  Specify enabled the delete button only if there were no references to all of the descendants of a node.  I modified behavior to enable the delete button if a node did not have children and there were no references to it.

DESCRIPTION: modified the isOkToDelete method in BaseTreeBusRules.