ISSUE: BugID 495 (Tree by Parent branch)
PURPOSE: Disabled locks on trees but kept them on defs and def items.  To release a def item lock, one needs to close the tab.  If one opens a locked a def item, he or she will get a "Tool In Use" dialog with the "Cancel" and "Try Again" buttons.  If the locking user closes the tab, clicking on "Try Again" will not work.  The new user has to click "Cancel" and then reopen the tree def.  I think this was the behavior before my changes, so whatever.

DESCRIPTION: Modified locking behaviors in BaseTreeTask.